+++ bin/opswatch
@@ -3,6 +3,5 @@
 GEM_ROOT = File.expand_path("../..", __FILE__)
 
 require 'opswatch'
-require 'pry'
 
 Opswatch::Main.start ARGV
