+++ test/opswatch_test.rb
@@ -4,8 +4,4 @@ class OpswatchTest < Minitest::Test
   def test_that_it_has_a_version_number
     refute_nil ::Opswatch::VERSION
   end
-
-  def test_it_does_something_useful
-    assert false
-  end
 end
