+++ test/test_helper.rb
@@ -1,4 +1,6 @@
 $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
+GEM_ROOT = File.expand_path("../..", __FILE__)
+
 require "opswatch"
 
 require "minitest/autorun"
