--- tests/test_helper.rb
@@ -1,2 +1,0 @@
-GEM_ROOT = File.expand_path("../..", __FILE__)
-require 'pry'
